Refactor EKFACLinearOperator into separate file
